+++ processing/src/main/java/org/apache/carbondata/processing/newflow/parser/impl/PrimitiveParserImpl.java
@@ -20,10 +20,10 @@ package org.apache.carbondata.processing.newflow.parser.impl;
 
 import org.apache.carbondata.processing.newflow.parser.GenericParser;
 
-public class PrimitiveParserImpl implements GenericParser<Object> {
+public class PrimitiveParserImpl implements GenericParser<String> {
 
   @Override
-  public Object parse(Object data) {
+  public String parse(String data) {
     return data;
   }
