+++ processing/src/main/java/org/apache/carbondata/processing/newflow/parser/impl/RowParserImpl.java
@@ -79,7 +79,7 @@ public class RowParserImpl implements RowParser {
   }
 
   @Override
-  public Object[] parseRow(Object[] row) {
+  public Object[] parseRow(String[] row) {
     // If number of columns are less in a row then create new array with same size of header.
     if (row.length < numberOfColumns) {
       String[] temp = new String[numberOfColumns];
@@ -88,7 +88,7 @@ public class RowParserImpl implements RowParser {
     }
     Object[] out = new Object[genericParsers.length];
     for (int i = 0; i < genericParsers.length; i++) {
-      Object obj = row[inputMapping[i]];
+      String obj = row[inputMapping[i]];
       out[outputMapping[i]] = genericParsers[i].parse(obj);
     }
     return out;
