+++ processing/src/main/java/org/apache/carbondata/processing/store/AbstractCarbonFactHandler.java
@@ -1,0 +1,1302 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.apache.carbondata.processing.store;
+
+import java.io.ByteArrayOutputStream;
+import java.io.DataOutputStream;
+import java.io.File;
+import java.io.IOException;
+import java.math.BigDecimal;
+import java.nio.ByteBuffer;
+import java.util.ArrayList;
+import java.util.BitSet;
+import java.util.List;
+import java.util.Map;
+import java.util.concurrent.*;
+
+import org.apache.carbondata.common.logging.LogService;
+import org.apache.carbondata.common.logging.LogServiceFactory;
+import org.apache.carbondata.core.carbon.datastore.block.SegmentProperties;
+import org.apache.carbondata.core.carbon.metadata.CarbonMetadata;
+import org.apache.carbondata.core.carbon.metadata.schema.table.CarbonTable;
+import org.apache.carbondata.core.carbon.metadata.schema.table.column.ColumnSchema;
+import org.apache.carbondata.core.constants.CarbonCommonConstants;
+import org.apache.carbondata.core.datastorage.store.columnar.BlockIndexerStorageForInt;
+import org.apache.carbondata.core.datastorage.store.columnar.BlockIndexerStorageForNoInvertedIndex;
+import org.apache.carbondata.core.datastorage.store.columnar.ColumnGroupModel;
+import org.apache.carbondata.core.datastorage.store.columnar.IndexStorage;
+import org.apache.carbondata.core.datastorage.store.compression.ValueCompressionModel;
+import org.apache.carbondata.core.datastorage.store.dataholder.CarbonWriteDataHolder;
+import org.apache.carbondata.core.datastorage.util.StoreFactory;
+import org.apache.carbondata.core.keygenerator.KeyGenException;
+import org.apache.carbondata.core.keygenerator.KeyGenerator;
+import org.apache.carbondata.core.keygenerator.columnar.ColumnarSplitter;
+import org.apache.carbondata.core.keygenerator.columnar.impl.MultiDimKeyVarLengthEquiSplitGenerator;
+import org.apache.carbondata.core.keygenerator.factory.KeyGeneratorFactory;
+import org.apache.carbondata.core.util.CarbonProperties;
+import org.apache.carbondata.core.util.CarbonUtil;
+import org.apache.carbondata.core.util.DataTypeUtil;
+import org.apache.carbondata.core.util.ValueCompressionUtil;
+import org.apache.carbondata.processing.datatypes.GenericDataType;
+import org.apache.carbondata.processing.mdkeygen.file.FileManager;
+import org.apache.carbondata.processing.mdkeygen.file.IFileManagerComposite;
+import org.apache.carbondata.processing.store.colgroup.*;
+import org.apache.carbondata.processing.store.writer.CarbonFactDataWriter;
+import org.apache.carbondata.processing.store.writer.CarbonFactDataWriterImplForIntIndexAndAggBlock;
+import org.apache.carbondata.processing.store.writer.NodeHolder;
+import org.apache.carbondata.processing.store.writer.exception.CarbonDataWriterException;
+import org.apache.carbondata.processing.util.RemoveDictionaryUtil;
+
+import org.apache.spark.sql.types.Decimal;
+
+/**
+ * Fact data handler class to handle the fact data
+ */
+public abstract class AbstractCarbonFactHandler implements CarbonFactHandler {
+
+  /**
+   * LOGGER
+   */
+  private static final LogService LOGGER =
+      LogServiceFactory.getLogService(AbstractCarbonFactHandler.class.getName());
+  /**
+   * data writer
+   */
+  protected CarbonFactDataWriter dataWriter;
+  /**
+   * File manager
+   */
+  private IFileManagerComposite fileManager;
+
+  private Map<Integer, GenericDataType> complexIndexMap;
+  /**
+   * measure count
+   */
+  private int measureCount;
+  /**
+   * measure count
+   */
+  private int dimensionCount;
+  /**
+   * index of mdkey in incoming rows
+   */
+  private int mdKeyIndex;
+  /**
+   * blocklet size
+   */
+  protected int blockletSize;
+  /**
+   * mdkeyLength
+   */
+  private int mdkeyLength;
+  /**
+   * storeLocation
+   */
+  private String storeLocation;
+  /**
+   * databaseName
+   */
+  private String databaseName;
+  /**
+   * tableName
+   */
+  private String tableName;
+  /**
+   * table block size in MB
+   */
+  private int tableBlockSize;
+  /**
+   * otherMeasureIndex
+   */
+  private int[] otherMeasureIndex;
+  /**
+   * customMeasureIndex
+   */
+  private int[] customMeasureIndex;
+  /**
+   * dimLens
+   */
+  private int[] dimLens;
+  /**
+   * keyGenerator
+   */
+  private ColumnarSplitter columnarSplitter;
+  /**
+   * keyBlockHolder
+   */
+  protected CarbonKeyBlockHolder[] keyBlockHolder;
+  private boolean[] aggKeyBlock;
+  private boolean[] isNoDictionary;
+  private boolean isAggKeyBlock;
+  /**
+   * thread pool size to be used for block sort
+   */
+  private int thread_pool_size;
+  /**
+   * factLevels
+   */
+  private int[] surrogateIndex;
+  /**
+   * factKeyGenerator
+   */
+  private KeyGenerator factKeyGenerator;
+  /**
+   * aggKeyGenerator
+   */
+  private KeyGenerator keyGenerator;
+  private KeyGenerator[] complexKeyGenerator;
+  /**
+   * maskedByteRanges
+   */
+  private int[] maskedByte;
+  /**
+   * isDataWritingRequest
+   */
+  private int noDictionaryCount;
+  private ColumnGroupModel colGrpModel;
+  private int[] primitiveDimLens;
+  private char[] type;
+  private int[] completeDimLens;
+  private boolean[] isUseInvertedIndex;
+  /**
+   * data file attributes which will used for file construction
+   */
+  private CarbonDataFileAttributes carbonDataFileAttributes;
+
+  /**
+   * number of cores configured
+   */
+  protected int numberOfCores;
+
+  /**
+   * data directory location in carbon store path
+   */
+  private String carbonDataDirectoryPath;
+  /**
+   * no of complex dimensions
+   */
+  private int complexColCount;
+
+  /**
+   * no of column blocks
+   */
+  private int columnStoreCount;
+
+  /**
+   * column schema present in the table
+   */
+  private List<ColumnSchema> wrapperColumnSchemaList;
+
+  /**
+   * boolean to check whether dimension
+   * is of dictionary type or no dictionary time
+   */
+  private boolean[] dimensionType;
+
+  /**
+   * colCardinality for the merge case.
+   */
+  private int[] colCardinality;
+
+  /**
+   * Segment properties
+   */
+  private SegmentProperties segmentProperties;
+  /**
+   * flag to check for compaction flow
+   */
+  private boolean compactionFlow;
+
+  protected boolean useKettle;
+
+  protected boolean isSorted;
+  /**
+   * CarbonFactDataHandler constructor
+   */
+  public AbstractCarbonFactHandler(CarbonFactDataHandlerModel carbonFactDataHandlerModel) {
+    initParameters(carbonFactDataHandlerModel);
+    this.dimensionCount = carbonFactDataHandlerModel.getDimensionCount();
+    this.complexIndexMap = carbonFactDataHandlerModel.getComplexIndexMap();
+    this.primitiveDimLens = carbonFactDataHandlerModel.getPrimitiveDimLens();
+    this.useKettle = carbonFactDataHandlerModel.isUseKettle();
+    this.isAggKeyBlock = Boolean.parseBoolean(CarbonProperties.getInstance()
+        .getProperty(CarbonCommonConstants.AGGREAGATE_COLUMNAR_KEY_BLOCK,
+            CarbonCommonConstants.AGGREAGATE_COLUMNAR_KEY_BLOCK_DEFAULTVALUE));
+    this.carbonDataDirectoryPath = carbonFactDataHandlerModel.getCarbonDataDirectoryPath();
+    this.complexColCount = getComplexColsCount();
+    this.columnStoreCount =
+        this.colGrpModel.getNoOfColumnStore() + noDictionaryCount + complexColCount;
+
+    this.aggKeyBlock = new boolean[columnStoreCount];
+    this.isNoDictionary = new boolean[columnStoreCount];
+    this.isUseInvertedIndex = new boolean[columnStoreCount];
+    if (null != carbonFactDataHandlerModel.getIsUseInvertedIndex()) {
+      for (int i = 0; i < isUseInvertedIndex.length; i++) {
+        if (i < carbonFactDataHandlerModel.getIsUseInvertedIndex().length) {
+          isUseInvertedIndex[i] = carbonFactDataHandlerModel.getIsUseInvertedIndex()[i];
+        } else {
+          isUseInvertedIndex[i] = true;
+        }
+      }
+    }
+    int noDictStartIndex = this.colGrpModel.getNoOfColumnStore();
+    // setting true value for dims of high card
+    for (int i = 0; i < noDictionaryCount; i++) {
+      this.isNoDictionary[noDictStartIndex + i] = true;
+    }
+
+    if (isAggKeyBlock) {
+      int noDictionaryValue = Integer.parseInt(CarbonProperties.getInstance()
+          .getProperty(CarbonCommonConstants.HIGH_CARDINALITY_VALUE,
+              CarbonCommonConstants.HIGH_CARDINALITY_VALUE_DEFAULTVALUE));
+      int[] columnSplits = colGrpModel.getColumnSplit();
+      int dimCardinalityIndex = 0;
+      int aggIndex = 0;
+      for (int i = 0; i < columnSplits.length; i++) {
+        if (colGrpModel.isColumnar(i) && dimLens[dimCardinalityIndex] < noDictionaryValue) {
+          this.aggKeyBlock[aggIndex++] = true;
+          continue;
+        }
+        dimCardinalityIndex += columnSplits[i];
+        aggIndex++;
+      }
+
+      if (dimensionCount < dimLens.length) {
+        int allColsCount = getColsCount(dimensionCount);
+        List<Boolean> aggKeyBlockWithComplex = new ArrayList<Boolean>(allColsCount);
+        for (int i = 0; i < dimensionCount; i++) {
+          GenericDataType complexDataType = complexIndexMap.get(i);
+          if (complexDataType != null) {
+            complexDataType.fillAggKeyBlock(aggKeyBlockWithComplex, this.aggKeyBlock);
+          } else {
+            aggKeyBlockWithComplex.add(this.aggKeyBlock[i]);
+          }
+        }
+        this.aggKeyBlock = new boolean[allColsCount];
+        for (int i = 0; i < allColsCount; i++) {
+          this.aggKeyBlock[i] = aggKeyBlockWithComplex.get(i);
+        }
+      }
+      aggKeyBlock = arrangeUniqueBlockType(aggKeyBlock);
+    }
+  }
+
+  private void initParameters(CarbonFactDataHandlerModel carbonFactDataHandlerModel) {
+    this.databaseName = carbonFactDataHandlerModel.getDatabaseName();
+    this.tableBlockSize = carbonFactDataHandlerModel.getBlockSizeInMB();
+    this.tableName = carbonFactDataHandlerModel.getTableName();
+    this.type = carbonFactDataHandlerModel.getAggType();
+    this.segmentProperties = carbonFactDataHandlerModel.getSegmentProperties();
+    this.wrapperColumnSchemaList = carbonFactDataHandlerModel.getWrapperColumnSchema();
+    this.colCardinality = carbonFactDataHandlerModel.getColCardinality();
+    this.storeLocation = carbonFactDataHandlerModel.getStoreLocation();
+    this.measureCount = carbonFactDataHandlerModel.getMeasureCount();
+    this.mdkeyLength = carbonFactDataHandlerModel.getMdKeyLength();
+    this.mdKeyIndex = carbonFactDataHandlerModel.getMdKeyIndex();
+    this.noDictionaryCount = carbonFactDataHandlerModel.getNoDictionaryCount();
+    this.colGrpModel = segmentProperties.getColumnGroupModel();
+    this.completeDimLens = carbonFactDataHandlerModel.getDimLens();
+    this.dimLens = this.segmentProperties.getDimColumnsCardinality();
+    this.carbonDataFileAttributes = carbonFactDataHandlerModel.getCarbonDataFileAttributes();
+    //TODO need to pass carbon table identifier to metadata
+    CarbonTable carbonTable = CarbonMetadata.getInstance()
+        .getCarbonTable(databaseName + CarbonCommonConstants.UNDERSCORE + tableName);
+    this.isSorted = carbonTable.isSort();
+    dimensionType =
+        CarbonUtil.identifyDimensionType(carbonTable.getDimensionByTableName(tableName));
+
+    this.compactionFlow = carbonFactDataHandlerModel.isCompactionFlow();
+    // in compaction flow the measure with decimal type will come as spark decimal.
+    // need to convert it to byte array.
+    if (compactionFlow) {
+      try {
+        numberOfCores = Integer.parseInt(CarbonProperties.getInstance()
+            .getProperty(CarbonCommonConstants.NUM_CORES_COMPACTING,
+                CarbonCommonConstants.NUM_CORES_DEFAULT_VAL));
+      } catch (NumberFormatException exc) {
+        LOGGER.error("Configured value for property " + CarbonCommonConstants.NUM_CORES_COMPACTING
+            + "is wrong.Falling back to the default value "
+            + CarbonCommonConstants.NUM_CORES_DEFAULT_VAL);
+        numberOfCores = Integer.parseInt(CarbonCommonConstants.NUM_CORES_DEFAULT_VAL);
+      }
+    } else {
+      try {
+        numberOfCores = Integer.parseInt(CarbonProperties.getInstance()
+            .getProperty(CarbonCommonConstants.NUM_CORES_LOADING,
+                CarbonCommonConstants.NUM_CORES_DEFAULT_VAL));
+      } catch (NumberFormatException exc) {
+        LOGGER.error("Configured value for property " + CarbonCommonConstants.NUM_CORES_LOADING
+            + "is wrong.Falling back to the default value "
+            + CarbonCommonConstants.NUM_CORES_DEFAULT_VAL);
+        numberOfCores = Integer.parseInt(CarbonCommonConstants.NUM_CORES_DEFAULT_VAL);
+      }
+    }
+
+    // get blocklet size
+    this.blockletSize = Integer.parseInt(CarbonProperties.getInstance()
+        .getProperty(CarbonCommonConstants.BLOCKLET_SIZE,
+            CarbonCommonConstants.BLOCKLET_SIZE_DEFAULT_VAL));
+  }
+
+  private boolean[] arrangeUniqueBlockType(boolean[] aggKeyBlock) {
+    int counter = 0;
+    boolean[] uniqueBlock = new boolean[aggKeyBlock.length];
+    for (int i = 0; i < dimensionType.length; i++) {
+      if (dimensionType[i]) {
+        uniqueBlock[i] = aggKeyBlock[counter++];
+      } else {
+        uniqueBlock[i] = false;
+      }
+    }
+    return uniqueBlock;
+  }
+
+  private void setComplexMapSurrogateIndex(int dimensionCount) {
+    int surrIndex = 0;
+    for (int i = 0; i < dimensionCount; i++) {
+      GenericDataType complexDataType = complexIndexMap.get(i);
+      if (complexDataType != null) {
+        List<GenericDataType> primitiveTypes = new ArrayList<GenericDataType>();
+        complexDataType.getAllPrimitiveChildren(primitiveTypes);
+        for (GenericDataType eachPrimitive : primitiveTypes) {
+          eachPrimitive.setSurrogateIndex(surrIndex++);
+        }
+      } else {
+        surrIndex++;
+      }
+    }
+  }
+
+  /**
+   * This method will be used to get and update the step properties which will
+   * required to run this step
+   *
+   * @throws CarbonDataWriterException
+   */
+  public void initialise() throws CarbonDataWriterException {
+    fileManager = new FileManager();
+    fileManager.setName(new File(this.storeLocation).getName());
+    setWritingConfiguration();
+  }
+
+  // TODO remove after kettle flow is removed
+  protected NodeHolder processDataRows(List<Object[]> dataRows) throws CarbonDataWriterException {
+    Object[] max = new Object[measureCount];
+    Object[] min = new Object[measureCount];
+    int[] decimal = new int[measureCount];
+    Object[] uniqueValue = new Object[measureCount];
+    // to store index of the measure columns which are null
+    BitSet[] nullValueIndexBitSet = getMeasureNullValueIndexBitSet(measureCount);
+    for (int i = 0; i < max.length; i++) {
+      if (type[i] == CarbonCommonConstants.BIG_INT_MEASURE) {
+        max[i] = Long.MIN_VALUE;
+      } else if (type[i] == CarbonCommonConstants.SUM_COUNT_VALUE_MEASURE) {
+        max[i] = -Double.MAX_VALUE;
+      } else if (type[i] == CarbonCommonConstants.BIG_DECIMAL_MEASURE) {
+        max[i] = new BigDecimal(0.0);
+      } else {
+        max[i] = 0.0;
+      }
+    }
+    for (int i = 0; i < min.length; i++) {
+      if (type[i] == CarbonCommonConstants.BIG_INT_MEASURE) {
+        min[i] = Long.MAX_VALUE;
+        uniqueValue[i] = Long.MIN_VALUE;
+      } else if (type[i] == CarbonCommonConstants.SUM_COUNT_VALUE_MEASURE) {
+        min[i] = Double.MAX_VALUE;
+        uniqueValue[i] = Double.MIN_VALUE;
+      } else if (type[i] == CarbonCommonConstants.BIG_DECIMAL_MEASURE) {
+        min[i] = new BigDecimal(Double.MAX_VALUE);
+        uniqueValue[i] = new BigDecimal(Double.MIN_VALUE);
+      } else {
+        min[i] = 0.0;
+        uniqueValue[i] = 0.0;
+      }
+    }
+    for (int i = 0; i < decimal.length; i++) {
+      decimal[i] = 0;
+    }
+
+    byte[] startKey = null;
+    byte[] endKey = null;
+    byte[] noDictStartKey = null;
+    byte[] noDictEndKey = null;
+    CarbonWriteDataHolder[] dataHolder = initialiseDataHolder(dataRows.size());
+    CarbonWriteDataHolder keyDataHolder = initialiseKeyBlockHolder(dataRows.size());
+    CarbonWriteDataHolder noDictionaryKeyDataHolder = null;
+    if ((noDictionaryCount + complexColCount) > 0) {
+      noDictionaryKeyDataHolder = initialiseKeyBlockHolder(dataRows.size());
+    }
+
+    for (int count = 0; count < dataRows.size(); count++) {
+      Object[] row = dataRows.get(count);
+      byte[] mdKey = (byte[]) row[this.mdKeyIndex];
+      byte[] noDictionaryKey = null;
+      if (noDictionaryCount > 0 || complexIndexMap.size() > 0) {
+        noDictionaryKey = (byte[]) row[this.mdKeyIndex - 1];
+      }
+      ByteBuffer byteBuffer = null;
+      byte[] b = null;
+      if (count == 0) {
+        startKey = mdKey;
+        noDictStartKey = noDictionaryKey;
+      }
+      endKey = mdKey;
+      noDictEndKey = noDictionaryKey;
+      // add to key store
+      if (mdKey.length > 0) {
+        keyDataHolder.setWritableByteArrayValueByIndex(count, mdKey);
+      }
+      // for storing the byte [] for high card.
+      if (noDictionaryCount > 0 || complexIndexMap.size() > 0) {
+        noDictionaryKeyDataHolder.setWritableByteArrayValueByIndex(count, noDictionaryKey);
+      }
+      //Add all columns to keyDataHolder
+      keyDataHolder.setWritableByteArrayValueByIndex(count, this.mdKeyIndex, row);
+      // CHECKSTYLE:OFF Approval No:Approval-351
+      for (int k = 0; k < otherMeasureIndex.length; k++) {
+        if (type[otherMeasureIndex[k]] == CarbonCommonConstants.BIG_INT_MEASURE) {
+          if (null == row[otherMeasureIndex[k]]) {
+            nullValueIndexBitSet[otherMeasureIndex[k]].set(count);
+            dataHolder[otherMeasureIndex[k]].setWritableLongValueByIndex(count, 0L);
+          } else {
+            dataHolder[otherMeasureIndex[k]]
+                .setWritableLongValueByIndex(count, row[otherMeasureIndex[k]]);
+          }
+        } else {
+          if (null == row[otherMeasureIndex[k]]) {
+            nullValueIndexBitSet[otherMeasureIndex[k]].set(count);
+            dataHolder[otherMeasureIndex[k]].setWritableDoubleValueByIndex(count, 0.0);
+          } else {
+            dataHolder[otherMeasureIndex[k]]
+                .setWritableDoubleValueByIndex(count, row[otherMeasureIndex[k]]);
+          }
+        }
+      }
+      calculateMaxMin(max, min, decimal, otherMeasureIndex, row);
+      for (int i = 0; i < customMeasureIndex.length; i++) {
+        if (null == row[customMeasureIndex[i]]
+            && type[customMeasureIndex[i]] == CarbonCommonConstants.BIG_DECIMAL_MEASURE) {
+          BigDecimal val = BigDecimal.valueOf(0);
+          b = DataTypeUtil.bigDecimalToByte(val);
+          nullValueIndexBitSet[customMeasureIndex[i]].set(count);
+        } else {
+          if (this.compactionFlow) {
+            BigDecimal bigDecimal = ((Decimal) row[customMeasureIndex[i]]).toJavaBigDecimal();
+            b = DataTypeUtil.bigDecimalToByte(bigDecimal);
+          } else {
+            b = (byte[]) row[customMeasureIndex[i]];
+          }
+        }
+        byteBuffer = ByteBuffer.allocate(b.length + CarbonCommonConstants.INT_SIZE_IN_BYTE);
+        byteBuffer.putInt(b.length);
+        byteBuffer.put(b);
+        byteBuffer.flip();
+        b = byteBuffer.array();
+        dataHolder[customMeasureIndex[i]].setWritableByteArrayValueByIndex(count, b);
+      }
+      calculateMaxMin(max, min, decimal, customMeasureIndex, row);
+    }
+    calculateUniqueValue(min, uniqueValue);
+    byte[][] byteArrayValues = keyDataHolder.getByteArrayValues().clone();
+    byte[][] noDictionaryValueHolder = null;
+    if ((noDictionaryCount + complexColCount) > 0) {
+      noDictionaryValueHolder = noDictionaryKeyDataHolder.getByteArrayValues();
+    }
+    ValueCompressionModel compressionModel = ValueCompressionUtil
+        .getValueCompressionModel(max, min, decimal, uniqueValue, type, new byte[max.length]);
+    byte[][] writableMeasureDataArray =
+        StoreFactory.createDataStore(compressionModel).getWritableMeasureDataArray(dataHolder)
+            .clone();
+    NodeHolder nodeHolder =
+        getNodeHolderObject(writableMeasureDataArray, byteArrayValues, dataRows.size(), startKey,
+            endKey, compressionModel, noDictionaryValueHolder, noDictStartKey, noDictEndKey);
+    nodeHolder.setMeasureNullValueIndex(nullValueIndexBitSet);
+    LOGGER.info("Number Of records processed: " + dataRows.size());
+    return nodeHolder;
+  }
+
+  protected NodeHolder processDataRowsWithOutKettle(List<Object[]> dataRows)
+      throws CarbonDataWriterException {
+    Object[] max = new Object[measureCount];
+    Object[] min = new Object[measureCount];
+    int[] decimal = new int[measureCount];
+    Object[] uniqueValue = new Object[measureCount];
+    // to store index of the measure columns which are null
+    BitSet[] nullValueIndexBitSet = getMeasureNullValueIndexBitSet(measureCount);
+    for (int i = 0; i < max.length; i++) {
+      if (type[i] == CarbonCommonConstants.BIG_INT_MEASURE) {
+        max[i] = Long.MIN_VALUE;
+      } else if (type[i] == CarbonCommonConstants.SUM_COUNT_VALUE_MEASURE) {
+        max[i] = -Double.MAX_VALUE;
+      } else if (type[i] == CarbonCommonConstants.BIG_DECIMAL_MEASURE) {
+        max[i] = new BigDecimal(0.0);
+      } else {
+        max[i] = 0.0;
+      }
+    }
+    for (int i = 0; i < min.length; i++) {
+      if (type[i] == CarbonCommonConstants.BIG_INT_MEASURE) {
+        min[i] = Long.MAX_VALUE;
+        uniqueValue[i] = Long.MIN_VALUE;
+      } else if (type[i] == CarbonCommonConstants.SUM_COUNT_VALUE_MEASURE) {
+        min[i] = Double.MAX_VALUE;
+        uniqueValue[i] = Double.MIN_VALUE;
+      } else if (type[i] == CarbonCommonConstants.BIG_DECIMAL_MEASURE) {
+        min[i] = new BigDecimal(Double.MAX_VALUE);
+        uniqueValue[i] = new BigDecimal(Double.MIN_VALUE);
+      } else {
+        min[i] = 0.0;
+        uniqueValue[i] = 0.0;
+      }
+    }
+    for (int i = 0; i < decimal.length; i++) {
+      decimal[i] = 0;
+    }
+
+    byte[] startKey = null;
+    byte[] endKey = null;
+    byte[][] noDictStartKey = null;
+    byte[][] noDictEndKey = null;
+    CarbonWriteDataHolder[] dataHolder = initialiseDataHolder(dataRows.size());
+    CarbonWriteDataHolder keyDataHolder = initialiseKeyBlockHolderWithOutKettle(dataRows.size());
+    CarbonWriteDataHolder noDictionaryKeyDataHolder = null;
+    if ((noDictionaryCount + complexColCount) > 0) {
+      noDictionaryKeyDataHolder = initialiseKeyBlockHolderForNonDictionary(dataRows.size());
+    }
+
+    for (int count = 0; count < dataRows.size(); count++) {
+      Object[] row = dataRows.get(count);
+      byte[] mdKey = (byte[]) row[this.mdKeyIndex];
+      byte[][] noDictionaryKey = null;
+      if (noDictionaryCount > 0 || complexIndexMap.size() > 0) {
+        noDictionaryKey = (byte[][]) row[this.mdKeyIndex - 1];
+      }
+      ByteBuffer byteBuffer = null;
+      byte[] b = null;
+      if (count == 0) {
+        startKey = mdKey;
+        noDictStartKey = noDictionaryKey;
+      }
+      endKey = mdKey;
+      noDictEndKey = noDictionaryKey;
+      // add to key store
+      if (mdKey.length > 0) {
+        keyDataHolder.setWritableByteArrayValueByIndex(count, mdKey);
+      }
+      // for storing the byte [] for high card.
+      if (noDictionaryCount > 0 || complexIndexMap.size() > 0) {
+        noDictionaryKeyDataHolder.setWritableNonDictByteArrayValueByIndex(count, noDictionaryKey);
+      }
+
+      for (int k = 0; k < otherMeasureIndex.length; k++) {
+        if (type[otherMeasureIndex[k]] == CarbonCommonConstants.BIG_INT_MEASURE) {
+          if (null == row[otherMeasureIndex[k]]) {
+            nullValueIndexBitSet[otherMeasureIndex[k]].set(count);
+            dataHolder[otherMeasureIndex[k]].setWritableLongValueByIndex(count, 0L);
+          } else {
+            dataHolder[otherMeasureIndex[k]]
+                .setWritableLongValueByIndex(count, row[otherMeasureIndex[k]]);
+          }
+        } else {
+          if (null == row[otherMeasureIndex[k]]) {
+            nullValueIndexBitSet[otherMeasureIndex[k]].set(count);
+            dataHolder[otherMeasureIndex[k]].setWritableDoubleValueByIndex(count, 0.0);
+          } else {
+            dataHolder[otherMeasureIndex[k]]
+                .setWritableDoubleValueByIndex(count, row[otherMeasureIndex[k]]);
+          }
+        }
+      }
+      calculateMaxMin(max, min, decimal, otherMeasureIndex, row);
+      for (int i = 0; i < customMeasureIndex.length; i++) {
+        if (null == row[customMeasureIndex[i]]
+            && type[customMeasureIndex[i]] == CarbonCommonConstants.BIG_DECIMAL_MEASURE) {
+          BigDecimal val = BigDecimal.valueOf(0);
+          b = DataTypeUtil.bigDecimalToByte(val);
+          nullValueIndexBitSet[customMeasureIndex[i]].set(count);
+        } else {
+          if (this.compactionFlow) {
+            BigDecimal bigDecimal = ((Decimal) row[customMeasureIndex[i]]).toJavaBigDecimal();
+            b = DataTypeUtil.bigDecimalToByte(bigDecimal);
+          } else {
+            b = (byte[]) row[customMeasureIndex[i]];
+          }
+        }
+        byteBuffer = ByteBuffer.allocate(b.length + CarbonCommonConstants.INT_SIZE_IN_BYTE);
+        byteBuffer.putInt(b.length);
+        byteBuffer.put(b);
+        byteBuffer.flip();
+        b = byteBuffer.array();
+        dataHolder[customMeasureIndex[i]].setWritableByteArrayValueByIndex(count, b);
+      }
+      calculateMaxMin(max, min, decimal, customMeasureIndex, row);
+    }
+    calculateUniqueValue(min, uniqueValue);
+    byte[][] byteArrayValues = keyDataHolder.getByteArrayValues().clone();
+    byte[][][] noDictionaryValueHolder = null;
+    if ((noDictionaryCount + complexColCount) > 0) {
+      noDictionaryValueHolder = noDictionaryKeyDataHolder.getNonDictByteArrayValues();
+    }
+    ValueCompressionModel compressionModel = ValueCompressionUtil
+        .getValueCompressionModel(max, min, decimal, uniqueValue, type, new byte[max.length]);
+    byte[][] writableMeasureDataArray =
+        StoreFactory.createDataStore(compressionModel).getWritableMeasureDataArray(dataHolder)
+            .clone();
+    NodeHolder nodeHolder =
+        getNodeHolderObjectWithOutKettle(writableMeasureDataArray, byteArrayValues, dataRows.size(),
+            startKey, endKey, compressionModel, noDictionaryValueHolder, noDictStartKey,
+            noDictEndKey);
+    nodeHolder.setMeasureNullValueIndex(nullValueIndexBitSet);
+    LOGGER.info("Number Of records processed: " + dataRows.size());
+    return nodeHolder;
+  }
+
+  // TODO remove after kettle flow is removed
+  private NodeHolder getNodeHolderObject(byte[][] dataHolderLocal, byte[][] byteArrayValues,
+      int entryCountLocal, byte[] startkeyLocal, byte[] endKeyLocal,
+      ValueCompressionModel compressionModel, byte[][] noDictionaryData,
+      byte[] noDictionaryStartKey, byte[] noDictionaryEndKey)
+      throws CarbonDataWriterException {
+    byte[][][] noDictionaryColumnsData = null;
+    List<ArrayList<byte[]>> colsAndValues = new ArrayList<ArrayList<byte[]>>();
+    int complexColCount = getComplexColsCount();
+
+    for (int i = 0; i < complexColCount; i++) {
+      colsAndValues.add(new ArrayList<byte[]>());
+    }
+    int noOfColumn = colGrpModel.getNoOfColumnStore();
+    DataHolder[] dataHolders = getDataHolders(noOfColumn, byteArrayValues.length);
+    for (int i = 0; i < byteArrayValues.length; i++) {
+      byte[][] splitKey = columnarSplitter.splitKey(byteArrayValues[i]);
+
+      for (int j = 0; j < splitKey.length; j++) {
+        dataHolders[j].addData(splitKey[j], i);
+      }
+    }
+    if (noDictionaryCount > 0 || complexIndexMap.size() > 0) {
+      noDictionaryColumnsData = new byte[noDictionaryCount][noDictionaryData.length][];
+      for (int i = 0; i < noDictionaryData.length; i++) {
+        int complexColumnIndex = primitiveDimLens.length + noDictionaryCount;
+        byte[][] splitKey = RemoveDictionaryUtil
+            .splitNoDictionaryKey(noDictionaryData[i], noDictionaryCount + complexIndexMap.size());
+
+        int complexTypeIndex = 0;
+        for (int j = 0; j < splitKey.length; j++) {
+          //nodictionary Columns
+          if (j < noDictionaryCount) {
+            noDictionaryColumnsData[j][i] = splitKey[j];
+          }
+          //complex types
+          else {
+            // Need to write columnar block from complex byte array
+            int index = complexColumnIndex - noDictionaryCount;
+            GenericDataType complexDataType = complexIndexMap.get(index);
+            complexColumnIndex++;
+            if (complexDataType != null) {
+              List<ArrayList<byte[]>> columnsArray = new ArrayList<ArrayList<byte[]>>();
+              for (int k = 0; k < complexDataType.getColsCount(); k++) {
+                columnsArray.add(new ArrayList<byte[]>());
+              }
+
+              try {
+                ByteBuffer complexDataWithoutBitPacking = ByteBuffer.wrap(splitKey[j]);
+                byte[] complexTypeData = new byte[complexDataWithoutBitPacking.getShort()];
+                complexDataWithoutBitPacking.get(complexTypeData);
+
+                ByteBuffer byteArrayInput = ByteBuffer.wrap(complexTypeData);
+                ByteArrayOutputStream byteArrayOutput = new ByteArrayOutputStream();
+                DataOutputStream dataOutputStream = new DataOutputStream(byteArrayOutput);
+                complexDataType
+                    .parseAndBitPack(byteArrayInput, dataOutputStream, this.complexKeyGenerator);
+                complexDataType.getColumnarDataForComplexType(columnsArray,
+                    ByteBuffer.wrap(byteArrayOutput.toByteArray()));
+                byteArrayOutput.close();
+              } catch (IOException e) {
+                throw new CarbonDataWriterException(
+                    "Problem while bit packing and writing complex datatype", e);
+              } catch (KeyGenException e) {
+                throw new CarbonDataWriterException(
+                    "Problem while bit packing and writing complex datatype", e);
+              }
+
+              for (ArrayList<byte[]> eachColumn : columnsArray) {
+                colsAndValues.get(complexTypeIndex++).addAll(eachColumn);
+              }
+            } else {
+              // This case not possible as ComplexType is the last columns
+            }
+          }
+        }
+      }
+    }
+    thread_pool_size = Integer.parseInt(CarbonProperties.getInstance()
+        .getProperty(CarbonCommonConstants.NUM_CORES_BLOCK_SORT,
+            CarbonCommonConstants.NUM_CORES_BLOCK_SORT_DEFAULT_VAL));
+    ExecutorService executorService = Executors.newFixedThreadPool(thread_pool_size);
+    List<Future<IndexStorage>> submit = new ArrayList<Future<IndexStorage>>(
+        primitiveDimLens.length + noDictionaryCount + complexColCount);
+    int i = 0;
+    int dictionaryColumnCount = -1;
+    int noDictionaryColumnCount = -1;
+    for (i = 0; i < dimensionType.length; i++) {
+      if (dimensionType[i]) {
+        dictionaryColumnCount++;
+        if (colGrpModel.isColumnar(dictionaryColumnCount)) {
+          submit.add(executorService
+              .submit(new BlockSortThread(i, dataHolders[dictionaryColumnCount].getData(),
+                  true, isUseInvertedIndex[i])));
+        } else {
+          submit.add(
+              executorService.submit(new ColGroupBlockStorage(dataHolders[dictionaryColumnCount])));
+        }
+      } else {
+        submit.add(executorService.submit(
+            new BlockSortThread(i, noDictionaryColumnsData[++noDictionaryColumnCount], false, true,
+                true, isUseInvertedIndex[i])));
+      }
+    }
+    for (int k = 0; k < complexColCount; k++) {
+      submit.add(executorService.submit(new BlockSortThread(i++,
+          colsAndValues.get(k).toArray(new byte[colsAndValues.get(k).size()][]), false, true)));
+    }
+    executorService.shutdown();
+    try {
+      executorService.awaitTermination(1, TimeUnit.DAYS);
+    } catch (InterruptedException e) {
+      LOGGER.error(e, e.getMessage());
+    }
+    IndexStorage[] blockStorage =
+        new IndexStorage[colGrpModel.getNoOfColumnStore() + noDictionaryCount + complexColCount];
+    try {
+      for (int k = 0; k < blockStorage.length; k++) {
+        blockStorage[k] = submit.get(k).get();
+      }
+    } catch (Exception e) {
+      LOGGER.error(e, e.getMessage());
+    }
+    return this.dataWriter
+        .buildDataNodeHolder(blockStorage, dataHolderLocal, entryCountLocal, startkeyLocal,
+            endKeyLocal, compressionModel, noDictionaryStartKey, noDictionaryEndKey);
+  }
+
+  private NodeHolder getNodeHolderObjectWithOutKettle(byte[][] dataHolderLocal,
+      byte[][] byteArrayValues, int entryCountLocal, byte[] startkeyLocal, byte[] endKeyLocal,
+      ValueCompressionModel compressionModel, byte[][][] noDictionaryData,
+      byte[][] noDictionaryStartKey, byte[][] noDictionaryEndKey)
+      throws CarbonDataWriterException {
+    byte[][][] noDictionaryColumnsData = null;
+    List<ArrayList<byte[]>> colsAndValues = new ArrayList<ArrayList<byte[]>>();
+    int complexColCount = getComplexColsCount();
+
+    for (int i = 0; i < complexColCount; i++) {
+      colsAndValues.add(new ArrayList<byte[]>());
+    }
+    int noOfColumn = colGrpModel.getNoOfColumnStore();
+    DataHolder[] dataHolders = getDataHolders(noOfColumn, byteArrayValues.length);
+    for (int i = 0; i < byteArrayValues.length; i++) {
+      byte[][] splitKey = columnarSplitter.splitKey(byteArrayValues[i]);
+
+      for (int j = 0; j < splitKey.length; j++) {
+        dataHolders[j].addData(splitKey[j], i);
+      }
+    }
+    if (noDictionaryCount > 0 || complexIndexMap.size() > 0) {
+      noDictionaryColumnsData = new byte[noDictionaryCount][noDictionaryData.length][];
+      for (int i = 0; i < noDictionaryData.length; i++) {
+        int complexColumnIndex = primitiveDimLens.length + noDictionaryCount;
+        byte[][] splitKey = noDictionaryData[i];
+
+        int complexTypeIndex = 0;
+        for (int j = 0; j < splitKey.length; j++) {
+          //nodictionary Columns
+          if (j < noDictionaryCount) {
+            int keyLength = splitKey[j].length;
+            byte[] newKey = new byte[keyLength + 2];
+            ByteBuffer buffer = ByteBuffer.wrap(newKey);
+            buffer.putShort((short)keyLength);
+            System.arraycopy(splitKey[j], 0, newKey, 2, keyLength);
+            noDictionaryColumnsData[j][i] = newKey;
+          }
+          //complex types
+          else {
+            // Need to write columnar block from complex byte array
+            int index = complexColumnIndex - noDictionaryCount;
+            GenericDataType complexDataType = complexIndexMap.get(index);
+            complexColumnIndex++;
+            if (complexDataType != null) {
+              List<ArrayList<byte[]>> columnsArray = new ArrayList<ArrayList<byte[]>>();
+              for (int k = 0; k < complexDataType.getColsCount(); k++) {
+                columnsArray.add(new ArrayList<byte[]>());
+              }
+
+              try {
+                ByteBuffer byteArrayInput = ByteBuffer.wrap(splitKey[j]);
+                ByteArrayOutputStream byteArrayOutput = new ByteArrayOutputStream();
+                DataOutputStream dataOutputStream = new DataOutputStream(byteArrayOutput);
+                complexDataType
+                    .parseAndBitPack(byteArrayInput, dataOutputStream, this.complexKeyGenerator);
+                complexDataType.getColumnarDataForComplexType(columnsArray,
+                    ByteBuffer.wrap(byteArrayOutput.toByteArray()));
+                byteArrayOutput.close();
+              } catch (IOException e) {
+                throw new CarbonDataWriterException(
+                    "Problem while bit packing and writing complex datatype", e);
+              } catch (KeyGenException e) {
+                throw new CarbonDataWriterException(
+                    "Problem while bit packing and writing complex datatype", e);
+              }
+
+              for (ArrayList<byte[]> eachColumn : columnsArray) {
+                colsAndValues.get(complexTypeIndex++).addAll(eachColumn);
+              }
+            } else {
+              // This case not possible as ComplexType is the last columns
+            }
+          }
+        }
+      }
+    }
+    thread_pool_size = Integer.parseInt(CarbonProperties.getInstance()
+        .getProperty(CarbonCommonConstants.NUM_CORES_BLOCK_SORT,
+            CarbonCommonConstants.NUM_CORES_BLOCK_SORT_DEFAULT_VAL));
+    ExecutorService executorService = Executors.newFixedThreadPool(thread_pool_size);
+    List<Future<IndexStorage>> submit = new ArrayList<Future<IndexStorage>>(
+        primitiveDimLens.length + noDictionaryCount + complexColCount);
+    int i = 0;
+    int dictionaryColumnCount = -1;
+    int noDictionaryColumnCount = -1;
+    for (i = 0; i < dimensionType.length; i++) {
+      if (dimensionType[i]) {
+        dictionaryColumnCount++;
+        if (colGrpModel.isColumnar(dictionaryColumnCount)) {
+          submit.add(executorService
+              .submit(new BlockSortThread(i, dataHolders[dictionaryColumnCount].getData(),
+                  this.isSorted, isUseInvertedIndex[i])));
+        } else {
+          submit.add(
+              executorService.submit(new ColGroupBlockStorage(dataHolders[dictionaryColumnCount])));
+        }
+      } else {
+        submit.add(executorService.submit(
+            new BlockSortThread(i, noDictionaryColumnsData[++noDictionaryColumnCount], false, true,
+                this.isSorted, isUseInvertedIndex[i])));
+      }
+    }
+    for (int k = 0; k < complexColCount; k++) {
+      submit.add(executorService.submit(new BlockSortThread(i++,
+          colsAndValues.get(k).toArray(new byte[colsAndValues.get(k).size()][]), false, true)));
+    }
+    executorService.shutdown();
+    try {
+      executorService.awaitTermination(1, TimeUnit.DAYS);
+    } catch (InterruptedException e) {
+      LOGGER.error(e, e.getMessage());
+    }
+    IndexStorage[] blockStorage =
+        new IndexStorage[colGrpModel.getNoOfColumnStore() + noDictionaryCount + complexColCount];
+    try {
+      for (int k = 0; k < blockStorage.length; k++) {
+        blockStorage[k] = submit.get(k).get();
+      }
+    } catch (Exception e) {
+      LOGGER.error(e, e.getMessage());
+    }
+    byte[] composedNonDictStartKey = null;
+    byte[] composedNonDictEndKey = null;
+    if (noDictionaryStartKey != null) {
+      composedNonDictStartKey =
+          RemoveDictionaryUtil.packByteBufferIntoSingleByteArray(noDictionaryStartKey);
+      composedNonDictEndKey =
+          RemoveDictionaryUtil.packByteBufferIntoSingleByteArray(noDictionaryEndKey);
+    }
+    return this.dataWriter
+        .buildDataNodeHolder(blockStorage, dataHolderLocal, entryCountLocal, startkeyLocal,
+            endKeyLocal, compressionModel, composedNonDictStartKey, composedNonDictEndKey);
+  }
+
+
+  /**
+   * DataHolder will have all row mdkey data
+   *
+   * @param noOfColumn : no of column participated in mdkey
+   * @param noOfRow    : total no of row
+   * @return : dataholder
+   */
+  private DataHolder[] getDataHolders(int noOfColumn, int noOfRow) {
+    DataHolder[] dataHolders = new DataHolder[noOfColumn];
+    int colGrpId = -1;
+    for (int colGrp = 0; colGrp < noOfColumn; colGrp++) {
+      if (colGrpModel.isColumnar(colGrp)) {
+        dataHolders[colGrp] = new ColumnDataHolder(noOfRow);
+      } else {
+        ColGroupMinMax colGrpMinMax = new ColGroupMinMax(segmentProperties, ++colGrpId);
+        dataHolders[colGrp] =
+            new ColGroupDataHolder(this.columnarSplitter.getBlockKeySize()[colGrp], noOfRow,
+                colGrpMinMax);
+      }
+    }
+    return dataHolders;
+  }
+
+  private byte[] getAggregateTableMdkey(byte[] maksedKey) throws CarbonDataWriterException {
+    long[] keyArray = this.factKeyGenerator.getKeyArray(maksedKey, maskedByte);
+
+    int[] aggSurrogateKey = new int[surrogateIndex.length];
+
+    for (int j = 0; j < aggSurrogateKey.length; j++) {
+      aggSurrogateKey[j] = (int) keyArray[surrogateIndex[j]];
+    }
+
+    try {
+      return keyGenerator.generateKey(aggSurrogateKey);
+    } catch (KeyGenException e) {
+      throw new CarbonDataWriterException("Problem while generating the mdkeyfor aggregate ", e);
+    }
+  }
+
+  private int getColsCount(int columnSplit) {
+    int count = 0;
+    for (int i = 0; i < columnSplit; i++) {
+      GenericDataType complexDataType = complexIndexMap.get(i);
+      if (complexDataType != null) {
+        count += complexDataType.getColsCount();
+      } else count++;
+    }
+    return count;
+  }
+
+  private int getComplexColsCount() {
+    int count = 0;
+    for (int i = 0; i < dimensionCount; i++) {
+      GenericDataType complexDataType = complexIndexMap.get(i);
+      if (complexDataType != null) {
+        count += complexDataType.getColsCount();
+      }
+    }
+    return count;
+  }
+
+  /**
+   * @param value
+   * @return it return no of value after decimal
+   */
+  private int getDecimalCount(double value) {
+    String strValue = Double.toString(value);
+    int integerPlaces = strValue.indexOf('.');
+    int decimalPlaces = 0;
+    if (-1 != integerPlaces) {
+      decimalPlaces = strValue.length() - integerPlaces - 1;
+    }
+    return decimalPlaces;
+  }
+  /**
+   * This method will be used to update the max value for each measure
+   */
+  private void calculateMaxMin(Object[] max, Object[] min, int[] decimal, int[] msrIndex,
+      Object[] row) {
+    // Update row level min max
+    for (int i = 0; i < msrIndex.length; i++) {
+      int count = msrIndex[i];
+      if (row[count] != null) {
+        if (type[count] == CarbonCommonConstants.SUM_COUNT_VALUE_MEASURE) {
+          double value = (double) row[count];
+          double maxVal = (double) max[count];
+          double minVal = (double) min[count];
+          max[count] = (maxVal > value ? max[count] : value);
+          min[count] = (minVal < value ? min[count] : value);
+          int num = getDecimalCount(value);
+          decimal[count] = (decimal[count] > num ? decimal[count] : num);
+        } else if (type[count] == CarbonCommonConstants.BIG_INT_MEASURE) {
+          long value = (long) row[count];
+          long maxVal = (long) max[count];
+          long minVal = (long) min[count];
+          max[count] = (maxVal > value ? max[count] : value);
+          min[count] = (minVal < value ? min[count] : value);
+          decimal[count] = 0;
+        } else if (type[count] == CarbonCommonConstants.BIG_DECIMAL_MEASURE) {
+          byte[] buff = null;
+          // in compaction flow the measure with decimal type will come as spark decimal.
+          // need to convert it to byte array.
+          if (this.compactionFlow) {
+            BigDecimal bigDecimal = ((Decimal) row[count]).toJavaBigDecimal();
+            buff = DataTypeUtil.bigDecimalToByte(bigDecimal);
+          } else {
+            buff = (byte[]) row[count];
+          }
+          BigDecimal value = DataTypeUtil.byteToBigDecimal(buff);
+          BigDecimal minVal = (BigDecimal) min[count];
+          min[count] = minVal.min(value);
+        }
+      }
+    }
+  }
+
+  /**
+   * This method will calculate the unique value which will be used as storage
+   * key for null values of measures
+   *
+   * @param minValue
+   * @param uniqueValue
+   */
+  private void calculateUniqueValue(Object[] minValue, Object[] uniqueValue) {
+    for (int i = 0; i < measureCount; i++) {
+      if (type[i] == CarbonCommonConstants.BIG_INT_MEASURE) {
+        uniqueValue[i] = (long) minValue[i] - 1;
+      } else if (type[i] == CarbonCommonConstants.BIG_DECIMAL_MEASURE) {
+        BigDecimal val = (BigDecimal) minValue[i];
+        uniqueValue[i] = (val.subtract(new BigDecimal(1.0)));
+      } else {
+        uniqueValue[i] = (double) minValue[i] - 1;
+      }
+    }
+  }
+
+  /**
+   * Below method will be to configure fact file writing configuration
+   *
+   * @throws CarbonDataWriterException
+   */
+  private void setWritingConfiguration() throws CarbonDataWriterException {
+    LOGGER.info("Blocklet Size: " + blockletSize);
+    int dimSet =
+        Integer.parseInt(CarbonCommonConstants.DIMENSION_SPLIT_VALUE_IN_COLUMNAR_DEFAULTVALUE);
+    // if atleast one dimension is present then initialize column splitter otherwise null
+    int noOfColStore = colGrpModel.getNoOfColumnStore();
+    int[] keyBlockSize = new int[noOfColStore + complexColCount];
+
+    if (dimLens.length > 0) {
+      //Using Variable length variable split generator
+      //This will help in splitting mdkey to columns. variable split is required because all
+      // columns which are part of
+      //row store will be in single column store
+      //e.g if {0,1,2,3,4,5} is dimension and {0,1,2) is row store dimension
+      //than below splitter will return column as {0,1,2}{3}{4}{5}
+      this.columnarSplitter = this.segmentProperties.getFixedLengthKeySplitter();
+      System.arraycopy(columnarSplitter.getBlockKeySize(), 0, keyBlockSize, 0, noOfColStore);
+      this.keyBlockHolder =
+          new CarbonKeyBlockHolder[this.columnarSplitter.getBlockKeySize().length];
+    } else {
+      this.keyBlockHolder = new CarbonKeyBlockHolder[0];
+    }
+    this.complexKeyGenerator = new KeyGenerator[completeDimLens.length];
+    for (int i = 0; i < completeDimLens.length; i++) {
+      complexKeyGenerator[i] =
+          KeyGeneratorFactory.getKeyGenerator(new int[] { completeDimLens[i] });
+    }
+
+    for (int i = 0; i < keyBlockHolder.length; i++) {
+      this.keyBlockHolder[i] = new CarbonKeyBlockHolder(blockletSize);
+      this.keyBlockHolder[i].resetCounter();
+    }
+
+    // agg type
+    List<Integer> otherMeasureIndexList =
+        new ArrayList<Integer>(CarbonCommonConstants.DEFAULT_COLLECTION_SIZE);
+    List<Integer> customMeasureIndexList =
+        new ArrayList<Integer>(CarbonCommonConstants.DEFAULT_COLLECTION_SIZE);
+    for (int j = 0; j < type.length; j++) {
+      if (type[j] != CarbonCommonConstants.BYTE_VALUE_MEASURE
+          && type[j] != CarbonCommonConstants.BIG_DECIMAL_MEASURE) {
+        otherMeasureIndexList.add(j);
+      } else {
+        customMeasureIndexList.add(j);
+      }
+    }
+    otherMeasureIndex = new int[otherMeasureIndexList.size()];
+    customMeasureIndex = new int[customMeasureIndexList.size()];
+    for (int i = 0; i < otherMeasureIndex.length; i++) {
+      otherMeasureIndex[i] = otherMeasureIndexList.get(i);
+    }
+    for (int i = 0; i < customMeasureIndex.length; i++) {
+      customMeasureIndex[i] = customMeasureIndexList.get(i);
+    }
+    setComplexMapSurrogateIndex(this.dimensionCount);
+    int[] blockKeySize = getBlockKeySizeWithComplexTypes(new MultiDimKeyVarLengthEquiSplitGenerator(
+        CarbonUtil.getIncrementedCardinalityFullyFilled(completeDimLens.clone()), (byte) dimSet)
+        .getBlockKeySize());
+    System.arraycopy(blockKeySize, noOfColStore, keyBlockSize, noOfColStore,
+        blockKeySize.length - noOfColStore);
+    this.dataWriter =
+        getFactDataWriter(this.storeLocation, this.measureCount, this.mdkeyLength, this.tableName,
+            fileManager, keyBlockSize);
+    this.dataWriter.setIsNoDictionary(isNoDictionary);
+    // initialize the channel;
+    this.dataWriter.initializeWriter();
+    //initializeColGrpMinMax();
+  }
+
+  /**
+   * This method combines primitive dimensions with complex metadata columns
+   *
+   * @param primitiveBlockKeySize
+   * @return all dimensions cardinality including complex dimension metadata column
+   */
+  private int[] getBlockKeySizeWithComplexTypes(int[] primitiveBlockKeySize) {
+    int allColsCount = getComplexColsCount();
+    int[] blockKeySizeWithComplexTypes =
+        new int[this.colGrpModel.getNoOfColumnStore() + allColsCount];
+
+    List<Integer> blockKeySizeWithComplex =
+        new ArrayList<Integer>(blockKeySizeWithComplexTypes.length);
+    for (int i = 0; i < this.dimensionCount; i++) {
+      GenericDataType complexDataType = complexIndexMap.get(i);
+      if (complexDataType != null) {
+        complexDataType.fillBlockKeySize(blockKeySizeWithComplex, primitiveBlockKeySize);
+      } else {
+        blockKeySizeWithComplex.add(primitiveBlockKeySize[i]);
+      }
+    }
+    for (int i = 0; i < blockKeySizeWithComplexTypes.length; i++) {
+      blockKeySizeWithComplexTypes[i] = blockKeySizeWithComplex.get(i);
+    }
+
+    return blockKeySizeWithComplexTypes;
+  }
+
+  // TODO Remove after kettle flow got removed.
+  private CarbonWriteDataHolder initialiseKeyBlockHolder(int size) {
+    CarbonWriteDataHolder keyDataHolder = new CarbonWriteDataHolder();
+    keyDataHolder.initialiseByteArrayValues(size);
+    return keyDataHolder;
+  }
+
+  private CarbonWriteDataHolder initialiseKeyBlockHolderWithOutKettle(int size) {
+    CarbonWriteDataHolder keyDataHolder = new CarbonWriteDataHolder();
+    keyDataHolder.initialiseByteArrayValuesWithOutKettle(size);
+    return keyDataHolder;
+  }
+
+  private CarbonWriteDataHolder initialiseKeyBlockHolderForNonDictionary(int size) {
+    CarbonWriteDataHolder keyDataHolder = new CarbonWriteDataHolder();
+    keyDataHolder.initialiseByteArrayValuesForNonDictionary(size);
+    return keyDataHolder;
+  }
+
+  private CarbonWriteDataHolder[] initialiseDataHolder(int size) {
+    CarbonWriteDataHolder[] dataHolder = new CarbonWriteDataHolder[this.measureCount];
+    for (int i = 0; i < otherMeasureIndex.length; i++) {
+      dataHolder[otherMeasureIndex[i]] = new CarbonWriteDataHolder();
+      if (type[otherMeasureIndex[i]] == CarbonCommonConstants.BIG_INT_MEASURE) {
+        dataHolder[otherMeasureIndex[i]].initialiseLongValues(size);
+      } else {
+        dataHolder[otherMeasureIndex[i]].initialiseDoubleValues(size);
+      }
+    }
+    for (int i = 0; i < customMeasureIndex.length; i++) {
+      dataHolder[customMeasureIndex[i]] = new CarbonWriteDataHolder();
+      dataHolder[customMeasureIndex[i]].initialiseByteArrayValues(size);
+    }
+    return dataHolder;
+  }
+
+  /**
+   * Below method will be used to get the bit set array for
+   * all the measure, which will store the indexes which are null
+   *
+   * @param measureCount
+   * @return bit set to store null value index
+   */
+  private BitSet[] getMeasureNullValueIndexBitSet(int measureCount) {
+    BitSet[] nullvalueIndexBitset = new BitSet[measureCount];
+    // creating a bit set for all the measure column
+    nullvalueIndexBitset = new BitSet[this.measureCount];
+    for (int i = 0; i < nullvalueIndexBitset.length; i++) {
+      // bitset size will be blocklet size
+      nullvalueIndexBitset[i] = new BitSet(this.blockletSize);
+    }
+    return nullvalueIndexBitset;
+  }
+
+  private CarbonFactDataWriter<?> getFactDataWriter(String storeLocation, int measureCount,
+      int mdKeyLength, String tableName, IFileManagerComposite fileManager, int[] keyBlockSize) {
+    return new CarbonFactDataWriterImplForIntIndexAndAggBlock(storeLocation, measureCount,
+        mdKeyLength, tableName, fileManager, keyBlockSize, aggKeyBlock, isComplexTypes(),
+        noDictionaryCount, carbonDataFileAttributes, databaseName, wrapperColumnSchemaList,
+        noDictionaryCount, dimensionType, carbonDataDirectoryPath, colCardinality,
+        segmentProperties, tableBlockSize);
+  }
+
+  private boolean[] isComplexTypes() {
+    int noOfColumn = colGrpModel.getNoOfColumnStore() + noDictionaryCount + complexIndexMap.size();
+    int allColsCount = getColsCount(noOfColumn);
+    boolean[] isComplexType = new boolean[allColsCount];
+
+    List<Boolean> complexTypesList = new ArrayList<Boolean>(allColsCount);
+    for (int i = 0; i < noOfColumn; i++) {
+      GenericDataType complexDataType = complexIndexMap.get(i - noDictionaryCount);
+      if (complexDataType != null) {
+        int count = complexDataType.getColsCount();
+        for (int j = 0; j < count; j++) {
+          complexTypesList.add(true);
+        }
+      } else {
+        complexTypesList.add(false);
+      }
+    }
+    for (int i = 0; i < allColsCount; i++) {
+      isComplexType[i] = complexTypesList.get(i);
+    }
+
+    return isComplexType;
+  }
+
+  private final class BlockSortThread implements Callable<IndexStorage> {
+    private int index;
+
+    private byte[][] data;
+    private boolean isSortRequired;
+    private boolean isCompressionReq;
+    private boolean isUseInvertedIndex;
+
+    private boolean isNoDictionary;
+
+    private BlockSortThread(int index, byte[][] data, boolean isSortRequired,
+        boolean isUseInvertedIndex) {
+      this.index = index;
+      this.data = data;
+      isCompressionReq = aggKeyBlock[this.index];
+      this.isSortRequired = isSortRequired;
+      this.isUseInvertedIndex = isUseInvertedIndex;
+    }
+
+    public BlockSortThread(int index, byte[][] data, boolean b, boolean isNoDictionary,
+        boolean isSortRequired, boolean isUseInvertedIndex) {
+      this.index = index;
+      this.data = data;
+      isCompressionReq = b;
+      this.isNoDictionary = isNoDictionary;
+      this.isSortRequired = isSortRequired;
+      this.isUseInvertedIndex = isUseInvertedIndex;
+    }
+
+    @Override public IndexStorage call() throws Exception {
+      if (isUseInvertedIndex) {
+        return new BlockIndexerStorageForInt(this.data, isCompressionReq, isNoDictionary,
+            isSortRequired);
+      } else {
+        return new BlockIndexerStorageForNoInvertedIndex(this.data, isCompressionReq,
+            isNoDictionary);
+      }
+
+    }
+
+  }
+}
